# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: coffee-api
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /beans:
    x-swagger-router-controller: beans
    post:
      description: Create new bean
      operationId: createBean
      parameters:
        - in: body
          name: body
          description: Bean object
          required: true
          schema:
            $ref: "#/definitions/Bean"
      responses:
        "201":
          description: Success
          schema:
            description: Success message
            type: string
        default:
          description: Error
          schema:
            description: Error message
            type: string
...
  /vessels:
    x-swagger-router-controller: vessels
    post:
      description: Create new vessel
      operationId: createVessel
      parameters:
        - in: body
          name: body
          description: Vessel object
          required: true
          schema:
            $ref: "#/definitions/Vessel"
      responses:
        "201":
          description: Success
          schema:
            description: Success message
            type: string
        default:
          description: Error
          schema:
            description: Error message
            type: string
  /vessels/{vesselName}:
    x-swagger-router-controller: vessels
    get:
      description: Returns info for vessel
      operationId: getVessel
      parameters:
        - in: path
          name: vesselName
          description: name of vessel
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Vessel"
        default:
          description: Error
          schema:
            description: Error message
            type: string
...
  /mixxy/master-lists:
    x-swagger-router-controller: mixxy/master_lists
    post:
      description: Create new master list
      operationId: createMixxyMasterList
      parameters:
        - in: body
          name: body
          description: master list object
          required: true
          schema:
            $ref: "#/definitions/MixxyMasterList"
      responses:
        "201":
          description: Success
          schema:
            description: Success message
            type: string
        default:
          description: Error
          schema:
            description: Error message
            type: string
    get:
      description: Returns all master lists
      operationId: getMixxyMasterLists
      responses:
        "200":
          description: Success
          schema:
            title: List of master lists
            type: array
            items:
              $ref: "#/definitions/MixxyMasterList"
        default:
          description: Error
          schema:
            description: Error message
            type: string
...
definitions:
...
  Bean:
    required:
      - sponsor_id
      - title
      - description
      - status
    properties:
      bean_id:
        type: string
        format: uuid
      sponsor_id:
        type: string
        format: uuid
      title:
        type: string
      description:
        type: string
      bean_link:
        type: string
      status:
        type: string
        enum: ["ACTIVE", "INACTIVE"]
...
  Vessel:
    required:
      - vessel_name
      - vessel_link
      - vessel_description
    properties:
      vessel_id:
        type: string
        format: uuid
      vessel_name:
        type: string
      vessel_link:
        type: string
      vessel_description:
        type: string
...
  MixxyMasterList:
    required:
      - name
      - short_description
      - long_description
      - image_link
    properties:
      campaign_id:
        type: string
        format: uuid
      tags:
        type: array
        items:
          type: string
      name:
        type: string
      short_description:
        type: string
      long_description:
        type: string
      image_link:
        type: string
      status:
        type: string
        enum: ["ACTIVE", "INACTIVE"]
...